Update patch set 1
Patch Set 1:

I would rather keep it as it is since there are scenarios where one can use the nli parameters only and leave out the raman parameters and vice versa. The misleading point here is that the RamanFiber class is more than a Fiber implementation that includes the raman profile. The distinction of the two classes was a safe step to be performed during the migration of the repo oopt-raman in oopt-gnpy, but at some point they should be merged or at least the new RamanFiber features that concern only the nli parameters should be included in the class Fiber.

Patch-set: 1
Reviewer: Gerrit User 1017876 <1017876@d5d70762-12d0-45a1-890d-524b12d3f735>