Heterogeneous channels for transmission-main-example
So far, when using GNPy as a tool and not a low-level, "please propagate
my signal" library, people could only use a homogeneous allocation of
channels. The would all be the same, and there would be no gaps. This
made it hard to use GNPy in brownfield scenarios where there might be
specific spectrum allocations. In short, GNPy really assumed the
worst-case planning scenario.

Let's start allowing people to provide their custom channel allocation.
In this patch it's only implemented for the transmission-main-example.
The other sample entry point we have, the path-requests-run, is geared
with spectrum allocation in response to users' requests. That would be a
bit tricky when operating on a preoccupied spectrum. Not exactly a
rocket science, but also not overly trivial, so I decided to implement
this as a proof-of-concept here.

Of course the ultimate approach (in my opinion, anyway) would be letting
people provide arbitrary "channel backfill" in an otherwise empty
network, and "be smart" about what we compute. I am not thrilled by the
idea of implementing this with our ad-hoc JSON parsing, so let's wait
for YANG before I get to that.

Change-Id: I12029c5ea02ffcb8a0f5ff9926bbd27a7a45c9ed