Update Pandas for better Python 3.8 compatibility
I kept wondering why installations on Python 3.8+ kept failing in the
CI, and I assumed that it was because the Pandas project somehow got
their homework wrong and did not provide a "wheel", which is a Python
thingy for prebuilt binaries. That's however not the case, and the
reason was simply that our Pandas version pin was too low, predating the
time when Python 3.8 was released.

Solve this by updating the Pandas requirement to the most recent one,
and now there's no need to mess with anything. Yay!

Thanks-to: https://github.com/mars-project/mars/pull/1332
Change-Id: Iceb31ae16aa911ebef67a938a3d2a524faad164f