CI: Prefer Python 3.8 and Fedora 32 over EL8
The first build was much faster, there was about 50s in the actual
pytest invocation time. Overall, the total time is about a minute
quicker, which is nice. I cannot qualify whether the improvement is due
to a 1g-1c -> 2g-2c VM flavor underneath, or if it's that newer Python
*and* updated compile options in Fedora -- but it's a nice improvement.
The downside is twice the cost of the VM hour.

Depends-on: https://review.gerrithub.io/c/Telecominfraproject/oopt-zuul-jobs/+/497370
See-also: https://fedoraproject.org/wiki/Changes/PythonNoSemanticInterpositionSpeedup
Change-Id: I6f5db61e9481170515b2f7ff4f3358fae5daa68e