YANG: Prepare for distributing YANG modules
I'm using a second-level module namespace (gnpy.yang) for the same
reasons as when shipping the example data via gnpy/example-data/ --
these will be used in the subsequent commits when we actually start
adding YANG models. There is also some code (not much now, a lot more in
future) for working with these models, and in future also for loading
actual data. These *could* be put into gnpy.tools.*, but I think it's
more straightforward to just keep them in the YANG namespace.

We require a pre-release of pyang because of a bug in Windows packaging
upstream.

Change-Id: Ic40738ddd8346429bde01e591d19fd2ce8cb687d