YANG: tests: validate our sample data against the YANG model
I'm using the yagnson library for this, and that library needs two
pieces of data as its inputs:

- a "YANG Module Library", which is usually a JSON description of all
available and activated YANG modules along its enabled features, etc,

- actual YANG models, typically specified as a list of filesystem paths
which hold them.

I generated that ietf-yanglib file via something like:

 $ python path/to/yangson/tools/python/mkylib.py \
     gnpy/yang/ext \
     gnpy/yang/tip \
     > gnpy/yang/yanglib.json`

When this adds support for `ietf-geo-location` in future, make sure to
edit the output so that it does not accidentally enable all of the
geolocation features (but that's for later, anyway). And we might
actually not end up doing that.

Change-Id: I51e342cd556ecc381ff0bf35df2bfa70f5f83ba8