YANG: support Raman simulations
This is once again "just the model", not an actual Python
implementation. What remains similar to the original JSON-based approach
is Raman-awareness on a per-fiber basis. However, instead of using a
special Fiber subclass, the model now provides just one kind of fiber,
and that fiber can be turned into a RamanFiber via one presence
container.

This requires several supporting pieces of information which are
mandatory as soon as there's any fiber with Raman-awareness activated:

- global simulation parameters (previously, this one went via
`sim_params.json`)
- the matching fiber in the equipment library has to specify the Raman
efficiency.

On a technical side of the YANG model, the augment direction might seem
counter-intuitive at first (I know it bit me). However, one cannot add a
`must` via an `augment`, and using a YANG deviation would feel
particularly silly given that I'm an author of both YANG models. Anyway,
having tried both ways I think it actually makes more sense to pull in
the simulation options from within the topology model.

I did the conversion via a simple Python script which -- I think -- will
become handy when we add a proper conversion from the "legacy JSON" into
the "new YANG-based JSON". Here's the gist of it:

 >>> with open('gnpy/example-data/eqpt_config.json', 'r') as f:
 ...  e=json.load(f)
 ...
 >>> d=e['RamanFiber'][0]['raman_efficiency']
 >>> print(',\n'.join(f'{{\n  "delta-frequency": "{int(k)}",\n  "cr": "{str(v)}"\n}}' for k,v in zip(d['frequency_offset'], d['cr'])))

FIXME: documentation needs to be written -- but it's definitely not a
regression compared to the current tree where it's also effectively
undocumented :(

FIXME: add more properties to the YANG model

Change-Id: I57e696db7a392372af8878bf923d2d0b77c0ebaa