Reading equipment catalog and the network from YANG files
To make sure that I get everything right, I built this code for
initializing the equipment library around the already existing JSON IO
loader. That is far from optimal because there is no type safety
whatsoever in these classes, and object properties are created in a
super ad-hoc manner at runtime. That is rather painful to work with
because there is no place anywhere in the code which would list all
properties that are *supposed* to be present.

Compared to that, the part which builds an actual network looks more
civilized -- in there we at least have some types to use, and these
types have the majority of their required options documented somewhere.
The distinction between "arguments passed to the constructor" and
"params" and "operation" is still something that I do not understand,
but hey, perhaps I'm missing something here.

FIXME: check some missing items for SI
FIXME: check some missing items for Span
FIXME: autodesign vs. "intentions" and "amplifier placement"

Change-Id: Ibbfd97a5a949cf107fd98484b19b24bf9f4ca3e9