Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> No sure I understand the try/except addition

I add the try/except scope to obtain a working version that pass all the test. In general, a simple if condition on the initialization of the simulation parameters dictionary should be included to be sure that, when no simulation parameters are given, the compute_nli method use the 'legacy' model of the nli estimation. The issue with the latter solution is that in some propagation test, even if no simulation parameters are given, some errors arise because the methods that require the simulation parameters are called anyway. This errors do not arise if one calls separately the test, but they arise when you run just pytest. I tried few attempts to fix this, but I think it would be more efficient if we can discuss about this issue.

Patch-set: 1