Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I'm confused. Firstly, this does not only apply to path_request_run, right? The same applies e.g. for transmission_main_example when you do a power sweep. Secondly, when I brought up a similar issue a long time ago (see #148) Jean-Luc replied that the fixed output power from a ROADM was intentional and supposed to model how ROADMs operate in practice. What has changed?

hello Jonas,
Thanks for this comment. The issue I noticed is that when using path_requests_run, there is the possibility for the user to specify a propagating input power into the line that may be different from the one specified in system design (SI/power_dbm). In this case the expected behaviour would be that the network is designed once and for all (all gain settings fixed) with this designed power, but that propagation of each services (full load with same power) would be made with the optional updated power.
I a way this means that the target_pch_out_db should be updated according to this new reference.
The current implementation correctly updates the target power for amplifiers but not for ROADMs, so that the first amp after ROADM (eg booster) changes its gain setting (and thus NF). I think this is not the desired behaviour, and this is why I proposed this patch. 
The overall way of working for ROADMs then remain the same as specified by Jean-Luc, the output power is set to a given power target, independently from the power history in the previous elements, so that roadm loss may change depending on the instance.

to illustrate this I have created a test with the detailed propagation for a path with 3 power settings (adding a print(el) in the propagate function) with and without this correction, the behaviour is the one that I expect. I Include this as additional files to a new patch so that you can review the change.

I think that this power handling will also be needed when we mix channels with different baudrate because we also will need to applied differentiated power target depending on the equalization strategy.

Patch-set: 1