Update patch set 4
Patch Set 4:

This topology model omits some information that we have in the current code, e.g. the location attributes city and region. Something like the city attribute is nice to have e.g. for grouping elements together when visualizing the topology. We should consider adding support for this in the YANG model.

Patch-set: 4