Fiber propagation of new Spectral Information. Nli solver methods coherently restructured.
Modification of the Fiber and the NliSolver in order to properly propagate the new definition of the spectral information taking advantage of the numpy array structures. This change include also a simplification of the FiberParameters and the NliParameters.

Additionally, in this change, previous issues regarding the loss_coef definition along the frequency are solved.

An additional test in test_science_utils.py verifies that the fiber propagation provides the correct values in case of a few cases of flex grid spectra.
Change-Id: Id71f36effba35fc3ed4bbf2481a3cf6566ccb51c