Update patch set 1
Patch Set 1:

> Honestly, this seems a bit backwards. Wouldn't it make more sense to have a single definition for both Fiber and RamanFiber in the equipment library if you anyway want to enforce the same parameter values?

That's how I want the YANG interface to work. In the network topology, you activate the Raman-aware engine via presence of the `/nw:networks/nw:network/nt:link/tip-topo:fiber/raman`. If it is there, then the code checks the equipment library to ensure that the matching `/tip-pe:fiber` fiber type has a `raman-efficiency` defined, and it goes ahead to instantiate an gnpy.core.elements.RamanFiber instead of g.c.elements.Fiber in the in-memory network topology. So in the YANG world, you will indeed have just a flat namespace holding input data for both g.c.equipment.Fiber an g.c.equipment.RamanFiber; both are represented via a `/tip-photonic-equipment:fiber` which might or might not have a `/tip-photonic-equipment:fiber/raman-efficiency`.

The problem with this approach is, however, that the code which we released in the past does not care whether an "SSMF" g.c.equipment.Fiber and "SSMF" g.c.equipment.RamanFiber define the same thing. Today, it's technically possible to have conflicting entries in the equipment library. For the YANG work I chose not to change the actual implementation of the equipment library, but to make sure that the code initializes the in-memory equipment library in a way which is compatible with what the user has intended. So, the code constructs an equipment['Fiber'] for each `/tip-photonic-equipment:fiber` instance, and it also creates an equipment['RamanFiber'] for each `/tip-photonic-equipment:fiber` instance which happens to have the presence container `/tip-photonic-equipment:fiber/raman-efficiency`. So far, so good. Without this patch, there's a possible problem when I'm converting the existing equipment library to its new YANG form. I'm creating a single `/tip-photonic-equipment:fiber` instance for each g.c.equipment.Fiber or g.c.equipment.RamanFiber. This needs checking that both of these legacy-JSON things share the same common parameters -- hence this patch.

Does that answer your concern? Should I improve the commit message by merging this info in there?

Patch-set: 1