Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > I understand the motivation but in my mind it would make more sense to change the implementation of the equipment library to match what you want.
> 
> How do you propose to handle existing user data in legacy JSON format which might contain a conflicting definition of a Fiber vs. RamanFiber with the same name, then?

Ok, that would still have to be handled in a similar way, I guess. In the longer run I think Fiber and RamanFiber should be combined into just Fiber, not only in the YANG interface but also in the equipment library implementation and in elements.py. But that is perhaps a separate issue.

Patch-set: 1