Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Please let me know if you know that this is going to break your use cases. Bonus points if you can also suggest what else could be done :).

for me this is not a hack: this is a NF modelling as for standard EDFA amp. It is less precise than the RAMAN feature, that can precisely determine the NF wrt fiber type and loss distribution, but once you have this NF the model gives the same result. We use is in Orange because it is fast and we can apply nf values provided by vendors. Besides, it is used by the autodesign to determine the hybrid amplifiers gain and NF.

Patch-set: 1
Reviewer: Gerrit User 1017751 <1017751@d5d70762-12d0-45a1-890d-524b12d3f735>