Update patch set 1
Patch Set 1: Code-Review+2

Ah, right. When I merged https://review.gerrithub.io/c/Telecominfraproject/oopt-gnpy/+/497382 I have not reviewed the XLS files, so I missed the fact that the file which actually defines the ROADM sheet is not used for end-to-end computation, just as a parser input. I think it would not hurt to add a test for this as well if you feel like doing that. I would prefer not to add it myself because I'm trying to stay as far away from the XLS code as possible 😎.

Patch-set: 1
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2