Update patch set 1
Patch Set 1:

> for me this is not a hack: this is a NF modelling as for standard EDFA amp. It is less precise than the RAMAN feature, that can precisely determine the NF wrt fiber type and loss distribution, but once you have this NF the model gives the same result.

Well, it only gives the same result for the ASE contribution, not for the nonlinear (NLI) contribution.

> We use is in Orange because it is fast and we can apply nf values provided by vendors.

Yes, I think the second part is at least as important as the speed. Even if the speed of the "true-Raman" code were greatly improved, it is still not really usable if you don't know the exact Raman pump wavelengths and power levels. If the vendor does not supply this info but only the effective NF, that is what you have to work with.

Patch-set: 1
Reviewer: Gerrit User 1017862 <1017862@d5d70762-12d0-45a1-890d-524b12d3f735>