Update patch set 1
Patch Set 1:

> I don't fully understand this proposal. There is nothing in the current YANG model stopping anyone from defining NF polynomials resulting in negative NF, right? So what would be the difference between "polynomial-NF" and this "fake-raman"?

I was told by a colleague of mine that there's a fundamental lower limit on an effective NF of an EDFA, it must be at least 3. If that's true, in my opinion it makes sense to introduce a sanity check and enforce that constraint upon each and every EDFA we model. On the other hand, it is apparently useful to define "another amplifier" which is not an EDFA via its "possibly negative NF". Rather than allowing a negative NF for all amplifiers (which is most likely a bug, right?), let's make this explicit in the model and describe this "not an EDFA" appropriately. I expect that to be less than 20 lines of code, including the YANG boilerplate.

> > The eqpt_config.json we ship only sets `raman = true` for the *dual stage* amps that use a Raman thing *internally*. The actual Raman stage is *not* marked as `raman = true` 😭.
> 
> I believe this is because the actual Raman stage is already marked "allowed_for_design": false and since the raman attribute is only used for selecting amplifiers by auto design, setting it to true would make no difference.

Yeah, I understand that it's harmless (it doesn't break anything), but this is an example of abstraction leaking if you ask me. We have an attribute that is called (and documented as) "this is a Raman amplifier", except that we do not set it for a Raman amplifier.

> Even if the speed of the "true-Raman" code were greatly improved, it is still not really usable if you don't know the exact Raman pump wavelengths and power levels. If the vendor does not supply this info but only the effective NF, that is what you have to work with.

This makes sense to me, so I'm going to call that new amplifier class something like "raman-approximation" and not "fake-raman".

Patch-set: 1