Bump all dependencies
I think it's "healthy" to sync up the list of minimal versions of our
dependencies with what the CI is actually checking, at least every now
and then. We're getting some variance via the list of different Python
interpreters, but still -- let's make sure that our dependencies match
what we're testing.

The only interesting case is scipy where the 1.6 release started
requiring Python 3.7. Let's stick with 1.5.x, which is the last version
supported on Python 3.6.

Change-Id: I4aa0a89e7c615f2095a951ff522a4838ef50178e