Introduce OpenROADM preamp and booster models
The NF calculated by the preamp model is compliant with the MW-MW noise
mask in the OpenROADM MSA spec. The booster is noise-free, which is
modeled by setting the NF to zero (-inf in dB units). This is obviously
unphysical but it is the simplest way to model the total noise
contribution from a ROADM, including preamp and booster, that is
compliant with the the OpenROADM MSA.

This also introduces two new EDFA type varieties, "mw_mw_preamp" and
"mw_mw_booster" in the equipment library. I would prefer to also change
the names of the existing "openroadm" type_def and
"standard"/"low_noise" type_variety, representing an OpenROADM
inline-amplifier, for better consistency but this probably needs to be
discussed first.

Signed-off-by: Jonas Mårtensson <jonas.martensson@ri.se>
Change-Id: I7344ff53623f166649efb389c95c04ff92718e35